# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Codex",
        "title": "API Service",
        "contact": {}
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/planguages/": {
            "post": {
                "description": "Adds Programming Language Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Add Programming Language",
                "parameters": [
                    {
                        "description": "New Programming Language",
                        "name": "newPLanguage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddPLanguageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates Programming Language Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Update Programming Language",
                "parameters": [
                    {
                        "description": "Update Programming Language",
                        "name": "updatePLanguage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePLanguageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/planguages/{id}": {
            "delete": {
                "description": "Delete Programming Languages from DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Delete Programming Language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Programming Language ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/rewards/": {
            "post": {
                "description": "Adds Reward Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Add Reward",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Reward Image File",
                        "name": "imageFile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "Description",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "rewardType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "sellerFee",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "symbol",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates Reward Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Update Reward",
                "parameters": [
                    {
                        "type": "string",
                        "name": "Description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "rewardType",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "symbol",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Course Image File",
                        "name": "imageFile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/rewards/{id}": {
            "delete": {
                "description": "Delete Rewards from DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Delete Reward",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/private/language/": {
            "get": {
                "description": "Retrieves all languages based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Language"
                ],
                "summary": "Get All Languages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Value",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/planguages/": {
            "get": {
                "description": "Retrieves all Programming languages based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Get All Programming Languages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Programming Language ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "languageID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Programming Language Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/planguages/{id}": {
            "get": {
                "description": "Retrieves spesific Programming languages based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Get One Programming Language By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Programming Language ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/rewards/": {
            "get": {
                "description": "Retrieves all rewards based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Get All Rewards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "rewardID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Type",
                        "name": "rewardType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/rewards/{id}": {
            "get": {
                "description": "Retrieves one reward.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Get Reward By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Reward Attribute Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Attribute Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddPLanguageDTO": {
            "type": "object",
            "required": [
                "fileExtention",
                "imagePath",
                "monacoEditor",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "fileExtention": {
                    "type": "string",
                    "maxLength": 30
                },
                "imagePath": {
                    "type": "string",
                    "maxLength": 60
                },
                "languageID": {
                    "type": "string"
                },
                "monacoEditor": {
                    "type": "string",
                    "maxLength": 30
                },
                "name": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "dto.UpdatePLanguageDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "fileExtention": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imagePath": {
                    "type": "string"
                },
                "languageID": {
                    "type": "string"
                },
                "monacoEditor": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "dataCount": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}